chore: windowsify runctx
